refactor: make constans package
